requirements.txt
numpy         
opencv-python 
Pillow        
pip           
setuptools    
UnrealCV      
wheel         
matplotlib
torch
